feat(cli): add watcherRunWithSpawn
Will probably the most used option
