+++ packages/cli/index.d.ts
@@ -1,5 +1,9 @@
 export type CliWatchOptions = import("./src/utils").CliWatchOptions;
 export { test, mainTestFn, newTestEvent } from "./src/testing/index.js";
 export { bench, mainBenchFn } from "./src/benchmarking/index.js";
-export { watcherKillProcess, watcherRun } from "./src/watcher.js";
+export {
+  watcherKillProcess,
+  watcherRun,
+  watcherRunWithSpawn,
+} from "./src/watcher.js";
 //# sourceMappingURL=index.d.ts.map
