+++ packages/cli/src/watcher.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="node" />
 /**
  * Kill a sub process with its own sub processes.
  * Uses tree-kill under the hood, to also exit any child processes of the provided process
@@ -32,5 +33,36 @@ export function watcherRun({
 }): {
   closeWatcher: () => Promise<void>;
 };
+/**
+ * Run watcher run & wrap around child process spawn.
+ * Makes sure the instance is fully killed, before starting up again.
+ *
+ * @param {Logger} logger
+ * @param {{
+ *   chokidarOptions: chokidar.WatchOptions,
+ *   hooks: {
+ *     onRestart: () => void,
+ *   }
+ * }} watcherOptions
+ * @param {{
+ *   cpArguments: Parameters<cpSpawn>
+ * }} spawnOptions
+ */
+export function watcherRunWithSpawn(
+  logger: Logger,
+  watcherOptions: {
+    chokidarOptions: chokidar.WatchOptions;
+    hooks: {
+      onRestart: () => void;
+    };
+  },
+  spawnOptions: {
+    cpArguments: [
+      command: string,
+      args: readonly string[],
+      options: import("child_process").SpawnOptions,
+    ];
+  },
+): void;
 import chokidar from "chokidar";
 //# sourceMappingURL=watcher.d.ts.map
